refactor: moved modules into different folders
